Refactored usd hook to resolver.py. Code improvements.
